OneDrive/מסמכים/m/requirements.txt
asgiref==3.8.1
asyncpg==0.30.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.4
Flask==3.1.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg-pool==3.2.4
psycopg2==2.9.10
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
